LayoutTests/platform/ipad/media/ios/video-volume-ios-quirk-expected.txt


Test 'volume' attribute

EXPECTED (video.volume == '1') OK
RUN(video.volume = 0.5)
EXPECTED (video.volume == '0.5') OK
RUN(video.volume = 0)
EXPECTED (video.volume == '0') OK
TEST(video.volume = 1.5) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.volume = -0.5) THROWS(DOMException.INDEX_SIZE_ERR) OK
LOG video.volume 100ms later: 0
END OF TEST
